% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoformat.R
\name{autoformat}
\alias{autoformat}
\title{Autoformat}
\usage{
autoformat(x, scheme = "KMBT", format = NULL)
}
\arguments{
\item{x}{Number or numeric vector}

\item{scheme}{Formatting scheme to apply. Default is "KMBT", meaning number will be shown as
trillions (T), billions (B), millions (M), thousands (K), or ones. In the case of any format
greater than ones, the number will be rounded to one decimal place.}

\item{format}{Format to apply (in absence of a scheme)}
}
\description{
Automatically formats a number or vector of numbers to a specified format
}
