% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrl_c.R
\name{ctrl_c}
\alias{ctrl_c}
\title{Control-C}
\usage{
ctrl_c(table, sep = "\\t", row.names = FALSE, col.names = TRUE)
}
\arguments{
\item{table}{A data table to be copied to the clipboard}
}
\description{
Copies a data frame to the clipboard (in Windows) in a format suitable for copy-and-paste into a spreadsheet program such as Microsoft Excel.
}
