% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_operators.R
\name{\%btwn\%}
\alias{\%btwn\%}
\title{Between operators}
\usage{
x \%btwn\% y
}
\description{
Between operators covering both inclusive and exclusive checks
}
\examples{
a <- c(0:10)
a \%btwn\% c(1,5) # inclusive between (1 and 5 included)
a \%_btwn_\% c(1,5) # exclusive between (1 and 5 not included)
}
