% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalency_operators.R
\name{\%==\%}
\alias{\%==\%}
\title{Equivalency operators}
\usage{
x \%==\% y
}
\description{
Equivalency operators that handle NAs, NULLs, and empty lists
  without throwing an error
}
\examples{
a <- NA
a == NA
a \%==\% NA

b <- NULL
b == NULL
b \%==\% NULL

x <- list()
x == list()
x \%==\% list()

a \%==\% b
a \%!=\% b
}
