% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in_operator.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not in operator}
\usage{
x \%!in\% y
}
\description{
Operator which tests whether vector or element x is not included in vector y. Equivalent to `!(x %in% y)`
}
\examples{
a <- c(1,2,3,4,5)
3 \%in\% a
3 \%!in\% a
6 \%in\% a
6 \%!in\% a

}
