% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{mae}
\alias{mae}
\title{Calculates mean absolute error}
\usage{
mae(y, pred)
}
\arguments{
\item{y}{Vector of observed values.}

\item{pred}{Vector of predicted values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculates mean absolute error from observed and predicted values.
}
\examples{
mae(1:10, 2:11)
}
