% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message0.R
\name{message0}
\alias{message0}
\title{Including variables in message, warning, and stop text.}
\usage{
message0(...)
}
\description{
Convenience function that replaces \code{message(paste0("This is ", x))}. Also available for \code{warning()} and \code{stop()}.
}
