% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefTables.R
\name{predict.coefTable}
\alias{predict.coefTable}
\title{Predicted values based on \code{coefTable} object.}
\usage{
\method{predict}{coefTable}(object, newdata, trafo = I, ...)
}
\arguments{
\item{object}{Object of class coefTable.}

\item{newdata}{New data.}

\item{trafo}{Transformation to be applied to resulting predictions, e.g. \code{exp}.}

\item{...}{Further arguments passed to \code{trafo}.}
}
\value{
Vector of predictions.
}
\description{
Takes a data frame and replaces randomly part of the values by missing values.
}
\examples{
CO2[["conc"]] <- factor(CO2[["conc"]])
fit <- lm(uptake ~ Type:Treatment + conc:Treatment + Type:conc, data = CO2)
coefT <- coefTables(fit, rebase = TRUE)
predict(fit, head(CO2))
predict(coefT, head(CO2))
}
