% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{rmse}
\alias{rmse}
\title{Calculates root mean squared error of prediction}
\usage{
rmse(y, pred)
}
\arguments{
\item{y}{Vector of observed values.}

\item{pred}{Vector of predicted values.}
}
\value{
A numeric vector of length one.
}
\description{
Calculates root mean squared error from observed and predicted values.
}
\examples{
rmse(1:10, 2:11)
}
