% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_load.R
\name{try_load}
\alias{try_load}
\title{Try to load a file before parsing an expression}
\usage{
try_load(filename, expr, save = FALSE, ...)
}
\arguments{
\item{filename}{Name of the file to try to load. Should be in RDS format.}

\item{expr}{Expression to run if the file doesn't exist.}

\item{save}{After expression is run, save the result as \code{filename}? Default is \code{FALSE}.}

\item{...}{Additional arguments to pass to \code{saveRDS} if \code{save == TRUE}.}
}
\description{
Checks whether a file exists, and if so loads it. If the file
doesn't exist, it runs an expression and optionally saves it
to that filename. Useful for scripts that take a long time to
run and where you want to save progress along the way.
}
\examples{
# Load setosa_glm.RDS if it exists, otherwise create and save it with this
# script
setosa_glm <- try_load("setosa_glm.RDS", save = TRUE,
  {
setosa <- iris[, 1:3]
setosa$setosa <- ifelse(iris$Species == "setosa", 1, 0)
setosa_glm <- glm(setosa ~ ., data = setosa)
  })

}
