% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messagef.R
\name{messagef}
\alias{messagef}
\title{Using sprintf formatting variables in message, warning, and stop text.}
\usage{
messagef(...)
}
\description{
Using sprintf formatting variables in message, warning, and stop text.
}
